Added selectDistinct procedure
selectDistinct returns all unique entries in a given column.

WARNING: Not all columns work as of yet. If you need to get the unique entries of a column not yet implemented, ask Tristan to add the functionality or add it yourself by adding a new switch case in the sample.ts